Add colors and sizes to markers based on popularity of event
Right now, these are static numbers.
That is okay for now, but it would not work in a professional rollout - if I live in a small town, my event will never be considered bumpin.
Later, I will write an algorithm that queries for all current events in a city and grabs an array of their popularity, and then splits them up into fifths.