Modularize event index controller into separate services (map, event, marker) and add template for front end rendering.
Original map.js is split into eventsController, mapService, eventService and markerService. mapService acts as a combined service with eventService and markerService. It allows storing the map without destroying with the dom element when ngView changes the view. 
map.js is changed to map.js.erb for accessing rails' asset path. It now has several routes that redirect to different templates, which are all stored in the asset templates folder.

Seed file is modified to get the first 6 user to generate the events, which means for testing, at least 6 users needs to be created before running "rake db:seed".

Add angular-route.js and angular-resource.js for routing functionality dependency.